+++ src/meltano/cli/repl.py
@@ -17,7 +17,7 @@ def repl(engine_uri):
         'print("\\nBooting import Meltano REPL\\n")',
         "from meltano.core.project import Project",
         "from meltano.core.plugin import PluginRef, PluginInstall, Plugin",
-        "from meltano.core.plugin.setting import PluginSetting",
+        "from meltano.core.setting import Setting",
         "from meltano.core.plugin.settings_service import PluginSettingsService",
         "from meltano.core.config_service import ConfigService",
         "from meltano.core.db import project_engine",