+++ src/meltano/core/plugin/base.py
@@ -6,6 +6,7 @@ from collections import namedtuple
 from enum import Enum
 from typing import Optional, Iterable
 
+from meltano.core.setting_definition import SettingDefinition
 from meltano.core.behavior.hookable import HookObject
 from meltano.core.behavior.canonical import Canonical
 from meltano.core.behavior import NameEq
@@ -30,43 +31,6 @@ class Profile(NameEq, Canonical):
 
 
 Profile.DEFAULT = Profile(name="default", label="Default")
-
-
-class SettingDefinition(NameEq, Canonical):
-    def __init__(
-        self,
-        name: str = None,
-        env: str = None,
-        kind: str = None,
-        value=None,
-        label: str = None,
-        documentation: str = None,
-        description: str = None,
-        tooltip: str = None,
-        options: list = [],
-        oauth: dict = {},
-        placeholder: str = None,
-        protected: bool = None,
-        **attrs,
-    ):
-        super().__init__(
-            # Attributes will be listed in meltano.yml in this order:
-            name=name,
-            env=env,
-            kind=kind,
-            value=value,
-            label=label,
-            documentation=documentation,
-            description=description,
-            tooltip=tooltip,
-            options=options,
-            oauth=oauth,
-            placeholder=placeholder,
-            protected=protected,
-            **attrs,
-        )
-
-        self._verbatim.add("value")
 
 
 class PluginType(YAMLEnum):