+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+- [#2122](https://gitlab.com/meltano/meltano/-/issues/2122) Allow custom (`meltano.yml`-`config`-defined) settings to be overridden using environment
+- [#2123](https://gitlab.com/meltano/meltano/-/issues/2123) Show current values and their source in `meltano config <plugin> list`
+
 ### Changes
 
 - [#2102](https://gitlab.com/meltano/meltano/-/issues/2101) Improve `discovery.yml` incompatibility handling and error messages